refactor: validataion 추가
