+++ src/main/java/techcourse/fakebook/utils/validator/PartitialNameValidator.java
@@ -1,11 +1,11 @@
-package techcourse.fakebook.utils;
+package techcourse.fakebook.utils.validator;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
 import java.util.regex.Pattern;
 
 public class PartitialNameValidator implements ConstraintValidator<PartitialName, String> {
-    private static final Pattern PARTITIAL_NAME_PATTERN = Pattern.compile("[a-zA-Z가-힣]{1,10}");
+    private static final String PARTITIAL_NAME_PATTERN = "[a-zA-Z가-힣]{1,10}";
 
     @Override
     public void initialize(PartitialName constraintAnnotation) {
@@ -13,6 +13,6 @@ public class PartitialNameValidator implements ConstraintValidator<PartitialName
 
     @Override
     public boolean isValid(String name, ConstraintValidatorContext context) {
-        return PARTITIAL_NAME_PATTERN.matcher(name).find();
+        return PARTITIAL_NAME_PATTERN.matches(name);
     }
 }
